refactor(kernel): nest cas_* primitives under core/cas/ per §4
`cas_chunking.rs`, `cas_engine.rs`, `cas_remote.rs`,
`cas_transport.rs` sat at the kernel crate root since before the
`core/` re-org (PR `c219d8abb`). KERNEL-ARCHITECTURE.md §4 says
kernel primitives live in `core/`, and the existing comment in
`abc/mod.rs` already refers to `CASEngine` "via `core/` modules"
— code lagged the doc.

Move the four files into a new `crate::core::cas` module:

* `cas_chunking.rs`  → `core/cas/chunking.rs`
* `cas_engine.rs`    → `core/cas/engine.rs`
* `cas_remote.rs`    → `core/cas/remote.rs`
* `cas_transport.rs` → `core/cas/transport.rs`

Internal sibling references inside the moved files switch to
`super::` (intra-`cas`) form. The external in-tree
`crate::cas_engine::CASEngine` reference in
`abc/object_store.rs::as_cas` is rewritten to the canonical
`crate::core::cas::engine::CASEngine` path.

To keep external consumers (notably `backends::storage::cas_local`
and the `backends::transports::api::ai::*` connectors) working
without a coordinated update, `lib.rs` re-exposes the four
modules at the historical flat paths via `pub use core::cas::*
as cas_*` shims; `kernel::cas_engine::CASEngine` etc. continue to
resolve to the same item.

Verification against the seven audit criteria:

1. Simpler contract — one home for kernel primitives (`core/`)
   instead of two (root + `core/`).
2. Architecture correct — aligns with §4. The doc-justified
   "must stay in the kernel crate" argument is unchanged (these
   *are* still in the kernel crate); only the *intra-crate* home
   shifts to `core/`.
3. SSOT — each cas file lives in exactly one place; the flat
   `cas_*` paths are pure re-exports.
4. DRY — n/a (file moves).
5. Rust-first — already pure Rust; no language move.
6. Perf — file location has no runtime impact; re-exports
   inline through the type system at zero cost.
7. Systematic — root cause was "files weren't moved when `core/`
   was introduced"; the move resolves it at the source.

`cargo check --workspace --tests` clean. Pre-existing 4
`core::fdt::tests::*` failures on Windows are unchanged
(unrelated audit noise).
